24 refactor streamlit (#24)
* chore: Refactoring

* feat: Refactor eyfs with bbc activities app
