+++ src/genai/streamlit_pages/__init__.py
@@ -1,0 +1,3 @@
+from .eli3_page import eli3
+from .eyfs_kb_bbc_page import eyfs_kb_bbc
+from .eyfs_page import early_year_activity_plan
